Split the "home_page" feature into activate/deactivate/explode
